backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
rest_framework_simplejwt==0.0.2
Psycopg
rest_framework_simplejwt
yfinance